# Conflicts:
#	FoodWars/Headers/GameECS/Components/DamageableComponent.h
#	FoodWars/Headers/GameECS/Components/PlayerComponent.h
#	FoodWars/Headers/StateMachine/Misc/Weapon.h
#	FoodWars/Sources/GameECS/Components/DamageableComponent.cpp
#	FoodWars/Sources/GameECS/Components/DamagingComponent.cpp
#	FoodWars/Sources/GameECS/Components/PlayerComponent.cpp
#	FoodWars/Sources/GameECS/Systems/StorageSystem.cpp
#	FoodWars/Sources/LevelLoader.cpp
#	FoodWars/Sources/StateMachine/Misc/Weapon.cpp
#	TonicEngine/Facades/Storage/StorageFacade.cpp
#	TonicEngine/Headers/Visual/Colour.h
#	TonicEngine/Sources/Visual/Colour.cpp

+++ FoodWars/Sources/GameECS/Systems/DamageableSystem.cpp
@@ -1,11 +1,15 @@
 #include "../../../Headers/GameECS/Systems/DamageableSystem.h"
 #include "../../../Headers/GameECS/Components/DamagingComponent.h"
+#include "../../../Headers/GameECS/Components/PlayerComponent.h"
+#include "../../../Headers/GameECS/Components/TurnComponent.h"
+#include <cmath>
 
 DamageableSystem::DamageableSystem(EntityManager &entityManager, AudioFacade& audioFacade,
                                    IObservable<CollisionEvent>& collisionEventObservable) :
         _entityManager{&entityManager},
         _audioFacade{&audioFacade},
-        CollisionEventHandler(collisionEventObservable) {
+        CollisionEventHandler(collisionEventObservable),
+        _damageCalculator{DamageCalculator{}}{
 }
 
 DamageableSystem::~DamageableSystem() = default;
@@ -14,7 +18,14 @@ void DamageableSystem::update(double deltaTime) {
     for (auto x : _entityManager->getAllEntitiesWithComponent<DamageableComponent>()) {
         if(!x.second->isAlive())
         {
-            _entityManager->removeEntity(x.first);
+            auto* playerComp = _entityManager->getComponentFromEntity<PlayerComponent>(x.first);
+            if(playerComp == nullptr) {
+                _entityManager->removeEntity(x.first);
+            }
+            //We've killed a player!
+            else{
+                playerComp->setIsAlive(false);
+            }
         }
     }
 }
@@ -28,11 +39,40 @@ bool DamageableSystem::canHandle(const CollisionEvent &collisionEvent) {
 
 void DamageableSystem::handleCollisionEvent(const CollisionEvent &collisionEvent) {
     auto projectile = _entityManager->getComponentFromEntity<DamageableComponent>(collisionEvent.getEntity());
+    auto projectileDamage = _entityManager->getComponentFromEntity<DamagingComponent>(collisionEvent.getEntity());
     projectile->destroy();
     auto target = _entityManager->getComponentFromEntity<DamageableComponent>(collisionEvent.getOtherEntity());
-    target->lowerHealth(
-            _entityManager->getComponentFromEntity<DamagingComponent>(collisionEvent.getEntity())->getDamage());
+    auto player = _entityManager->getComponentFromEntity<PlayerComponent>(collisionEvent.getOtherEntity());
+
+    int damage = 0;
+    if (player == nullptr) {
+        damage = projectileDamage->getDamage() - target->getResistance();
+    } else {
+        damage = _damageCalculator.calculateDamage(*projectileDamage, *target, *player);
+    }
+    if (damage > 0) target->lowerHealth(damage);
+
     _audioFacade->playEffect("damage");
 
     std::cout << "currentHP: " << target->getHealth() << std::endl;
+
+    // Default point increase/decrease
+    int iPoints = 10;
+    std::map<int, PlayerComponent*> players = _entityManager->getAllEntitiesWithComponent<PlayerComponent>();
+    for (auto const& player : _entityManager->getAllEntitiesWithComponent<PlayerComponent>())
+    {
+        if (player.first == collisionEvent.getOtherEntity())
+        {
+            player.second->addScore(-iPoints);
+            iPoints = damage;
+        }
+    }
+
+    for (auto const& turn : _entityManager->getAllEntitiesWithComponent<TurnComponent>())
+    {
+        if (turn.second->isMyTurn())
+        {
+            players[turn.first]->addScore(iPoints);
+        }
+    }
 }
